[MIG] operating_unit: Migration to 17.0
