+++ operating_unit/__manifest__.py
@@ -6,7 +6,7 @@
     "name": "Operating Unit",
     "summary": "An operating unit (OU) is an organizational entity part of a "
     "company",
-    "version": "16.0.1.0.2",
+    "version": "17.0.1.0.0",
     "author": "ForgeFlow, "
     "Serpent Consulting Services Pvt. Ltd., "
     "Odoo Community Association (OCA)",
