+++ operating_unit/data/operating_unit_data.xml
@@ -8,10 +8,16 @@
     </record>
     <record model="res.users" id="base.user_root">
         <field name="default_operating_unit_id" ref="main_operating_unit" />
-        <field name="operating_unit_ids" eval="[(4, ref('main_operating_unit'))]" />
+        <field
+            name="assigned_operating_unit_ids"
+            eval="[(4, ref('main_operating_unit'))]"
+        />
     </record>
     <record model="res.users" id="base.user_admin">
         <field name="default_operating_unit_id" ref="main_operating_unit" />
-        <field name="operating_unit_ids" eval="[(4, ref('main_operating_unit'))]" />
+        <field
+            name="assigned_operating_unit_ids"
+            eval="[(4, ref('main_operating_unit'))]"
+        />
     </record>
 </odoo>
