+++ operating_unit/models/__init__.py
@@ -1,2 +1,3 @@
+from . import ir_rule
 from . import operating_unit
 from . import res_users
