+++ operating_unit/models/operating_unit.py
@@ -41,22 +41,18 @@ class OperatingUnit(models.Model):
         ),
     ]
 
-    def name_get(self):
-        res = []
+    @api.depends("name", "code")
+    def _compute_display_name(self):
         for ou in self:
-            name = ou.name
-            if ou.code:
-                name = f"[{ou.code}] {name}"
-            res.append((ou.id, name))
-        return res
+            ou.display_name = f"[{ou.code}] {ou.name}"
 
     @api.model_create_multi
     def create(self, vals_list):
         res = super().create(vals_list)
         res.write({"user_ids": [fields.Command.link(self.env.user.id)]})
-        self.clear_caches()
+        self.env.registry.clear_cache()
         return res
 
     def write(self, vals):
-        self.clear_caches()
+        self.env.registry.clear_cache()
         return super().write(vals)
