+++ operating_unit/models/res_users.py
@@ -8,8 +8,32 @@ from odoo import api, fields, models
 class ResUsers(models.Model):
     _inherit = "res.users"
 
+    operating_unit_ids = fields.One2many(
+        comodel_name="operating.unit",
+        compute="_compute_operating_unit_ids",
+        inverse="_inverse_operating_unit_ids",
+        string="Allowed Operating Units",
+        compute_sudo=True,
+    )
+
+    assigned_operating_unit_ids = fields.Many2many(
+        comodel_name="operating.unit",
+        relation="operating_unit_users_rel",
+        column1="user_id",
+        column2="operating_unit_id",
+        string="Operating Units",
+        default=lambda self: self._default_operating_unit(),
+    )
+
+    default_operating_unit_id = fields.Many2one(
+        comodel_name="operating.unit",
+        string="Default Operating Unit",
+        default=lambda self: self._default_operating_unit(),
+        domain="[('company_id', '=', current_company_id)]",
+    )
+
     @api.model
-    def operating_unit_default_get(self, uid2=False):
+    def _get_default_operating_unit(self, uid2=False):
         if not uid2:
             uid2 = self.env.user.id
         user = self.env["res.users"].browse(uid2)
@@ -29,66 +53,11 @@ class ResUsers(models.Model):
 
     @api.model
     def _default_operating_unit(self):
-        return self.operating_unit_default_get()
-
-    @api.model
-    def _default_operating_units(self):
-        return self._default_operating_unit()
-
-    operating_unit_ids = fields.One2many(
-        comodel_name="operating.unit",
-        compute="_compute_operating_unit_ids",
-        inverse="_inverse_operating_unit_ids",
-        string="Allowed Operating Units",
-        compute_sudo=True,
-    )
-
-    assigned_operating_unit_ids = fields.Many2many(
-        comodel_name="operating.unit",
-        relation="operating_unit_users_rel",
-        column1="user_id",
-        column2="operating_unit_id",
-        string="Operating Units",
-        default=lambda self: self._default_operating_units(),
-    )
-
-    default_operating_unit_id = fields.Many2one(
-        comodel_name="operating.unit",
-        string="Default Operating Unit",
-        default=lambda self: self._default_operating_unit(),
-        domain="[('company_id', '=', current_company_id)]",
-    )
-
-    @api.onchange("operating_unit_ids")
-    def _onchange_operating_unit_ids(self):
-        for record in self:
-            if (
-                record.default_operating_unit_id
-                and record.default_operating_unit_id
-                not in record.operating_unit_ids._origin
-            ):
-                record.default_operating_unit_id = False
-
-    @api.depends("groups_id", "assigned_operating_unit_ids")
-    def _compute_operating_unit_ids(self):
-        for user in self:
-            if user._origin.has_group("operating_unit.group_manager_operating_unit"):
-                dom = []
-                if self.env.context.get("allowed_company_ids"):
-                    dom = [
-                        "|",
-                        ("company_id", "=", False),
-                        ("company_id", "in", self.env.context["allowed_company_ids"]),
-                    ]
-                else:
-                    dom = []
-                user.operating_unit_ids = self.env["operating.unit"].sudo().search(dom)
-            else:
-                user.operating_unit_ids = user.assigned_operating_unit_ids
+        return self._get_default_operating_unit()
 
     @api.model
     def default_get(self, fields):
-        vals = super(ResUsers, self).default_get(fields)
+        vals = super().default_get(fields)
         if (
             self.env["ir.config_parameter"]
             .sudo()
@@ -102,7 +71,37 @@ class ResUsers(models.Model):
             vals["operating_unit_ids"] = [(6, 0, default_user.operating_unit_ids.ids)]
         return vals
 
+    @api.depends("groups_id", "assigned_operating_unit_ids")
+    def _compute_operating_unit_ids(self):
+        for user in self:
+            if user._origin.has_group("operating_unit.group_manager_operating_unit"):
+                if self.env.context.get("allowed_company_ids"):
+                    dom = [
+                        "|",
+                        ("company_id", "=", False),
+                        ("company_id", "in", self.env.context["allowed_company_ids"]),
+                    ]
+                else:
+                    dom = []
+
+                user.operating_unit_ids = self.env["operating.unit"].sudo().search(dom)
+            else:
+                user.operating_unit_ids = user.assigned_operating_unit_ids
+
     def _inverse_operating_unit_ids(self):
         for user in self:
             user.assigned_operating_unit_ids = user.operating_unit_ids
-        self.clear_caches()
+        self.env.registry.clear_cache()
+
+    @api.onchange("operating_unit_ids")
+    def _onchange_operating_unit_ids(self):
+        for record in self:
+            if (
+                record.default_operating_unit_id
+                and record.default_operating_unit_id
+                not in record.operating_unit_ids._origin
+            ):
+                record.default_operating_unit_id = False
+
+    def operating_units(self):
+        return self.env.user.operating_unit_ids
