+++ operating_unit/security/operating_unit_security.xml
@@ -27,7 +27,7 @@
     </record>
     <record id="ir_rule_operating_unit_allowed_operating_units" model="ir.rule">
         <field name="model_id" ref="model_operating_unit" />
-        <field name="domain_force">[('id','in',user.operating_unit_ids.ids)]</field>
+        <field name="domain_force">[('id','in',operating_unit_ids)]</field>
         <field name="name">Allowed operating units</field>
         <field eval="0" name="perm_unlink" />
         <field eval="0" name="perm_write" />
