+++ operating_unit/tests/test_operating_unit.py
@@ -2,71 +2,14 @@
 # License LGPL-3.0 or later (https://www.gnu.org/licenses/lgpl.html)
 
 from odoo.exceptions import AccessError
-from odoo.tests import common
+from odoo.tests import tagged
 from odoo.tests.common import Form
 
+from .common import OperatingUnitCommon
 
-class TestOperatingUnit(common.TransactionCase):
-    @classmethod
-    def setUpClass(cls):
-        super().setUpClass()
-        cls.res_users_model = cls.env["res.users"].with_context(
-            tracking_disable=True, no_reset_password=True
-        )
 
-        # Groups
-        cls.grp_ou_mngr = cls.env.ref("operating_unit.group_manager_operating_unit")
-        cls.grp_ou_multi = cls.env.ref("operating_unit.group_multi_operating_unit")
-        # Company
-        cls.company = cls.env.ref("base.main_company")
-        cls.company_2 = cls.env["res.company"].create({"name": "Second company"})
-        # Main Operating Unit
-        cls.ou1 = cls.env.ref("operating_unit.main_operating_unit")
-        # B2C Operating Unit
-        cls.b2c = cls.env.ref("operating_unit.b2c_operating_unit")
-        # B2B Operating Unit
-        cls.b2b = cls.env.ref("operating_unit.b2b_operating_unit")
-        # Create User 1 with Main OU
-        cls.user1 = cls._create_user("user_1", cls.grp_ou_mngr, cls.company, cls.ou1)
-        # Create User 2 with B2C OU
-        cls.user2 = cls._create_user("user_2", cls.grp_ou_multi, cls.company, cls.b2c)
-
-    @classmethod
-    def _create_user(cls, login, group, company, operating_units, context=None):
-        """Create a user."""
-        user = cls.res_users_model.create(
-            {
-                "name": "Test User",
-                "login": login,
-                "password": "demo",
-                "email": "test@yourcompany.com",
-                "company_id": company.id,
-                "company_ids": [(4, company.id)],
-                "operating_unit_ids": [(4, ou.id) for ou in operating_units],
-                "default_operating_unit_id": False,
-                "sel_groups_13_14": group.id,
-            }
-        )
-        return user
-
-    def _create_operating_unit(self, uid, name, code, company_id=None):
-        """Create Operating Unit"""
-        if company_id is None:
-            company_id = self.company
-        ou = (
-            self.env["operating.unit"]
-            .with_user(uid)
-            .create(
-                {
-                    "name": name,
-                    "code": code,
-                    "partner_id": company_id.partner_id.id,
-                    "company_id": company_id.id,
-                }
-            )
-        )
-        return ou
-
+@tagged("post_install", "-at_install")
+class TestOperatingUnit(OperatingUnitCommon):
     def test_01_operating_unit(self):
         # User 1 tries to create and modify an OU
         # Create
@@ -156,17 +99,17 @@ class TestOperatingUnit(common.TransactionCase):
 
     def test_03_operating_unit(self):
         """
-        The method operating_unit_default_get should not return
+        The method _get_default_operating_unit should not return
         operating units belonging to a company that is not active
         """
         self.assertEqual(
-            self.res_users_model.operating_unit_default_get(uid2=self.user1.id),
+            self.res_users_model._get_default_operating_unit(uid2=self.user1.id),
             self.ou1,
         )
         self.assertEqual(
             self.res_users_model.with_company(
                 self.company_2
-            ).operating_unit_default_get(uid2=self.user1.id),
+            )._get_default_operating_unit(uid2=self.user1.id),
             False,
         )
 
@@ -178,7 +121,7 @@ class TestOperatingUnit(common.TransactionCase):
         self.assertEqual(
             self.res_users_model.with_company(
                 self.company_2
-            ).operating_unit_default_get(uid2=self.user1.id),
+            )._get_default_operating_unit(uid2=self.user1.id),
             ou_company_2,
         )
